requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pymupdf==1.22.5
python-docx==0.8.11
pdfminer.six==20221105
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
